rearrange & rename code around integration type info
