+++ clients/admin-ui/src/features/integrations/integration-type-info/dynamoInfo.tsx
