--- clients/admin-ui/src/features/integrations/supportedIntegrations.ts
@@ -1,3 +1,0 @@
-const SUPPORTED_INTEGRATIONS = ["bigquery", "dynamodb", "scylla", "s3"];
-
-export default SUPPORTED_INTEGRATIONS;
